groupchat_app_src/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0.0
asyncmy
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
httpx
jinja2
python-multipart
chromadb
sentence-transformers
PyPDF2
python-docx
numpy<2
